[ui, refactor] Extract all dashboard cards as separate server components
