+++ app/src/app/dashboard/custom-activity-category-card.tsx
@@ -1,0 +1,23 @@
+import { createClient } from "@/lib/supabase/server";
+import { DashboardCard } from "@/app/dashboard/dashboard-card";
+import { Settings } from "lucide-react";
+
+export const CustomActivityCategoryCard = async () => {
+  const client = createClient();
+
+  const { count, error } = await client
+    .from("activity_category_available_custom")
+    .select("", { count: "exact" })
+    .limit(0);
+
+  await new Promise((resolve) => setTimeout(resolve, 2000));
+
+  return (
+    <DashboardCard
+      title="Custom Activity Category Codes"
+      icon={<Settings className="h-4" />}
+      text={count?.toString() ?? "-"}
+      failed={!!error}
+    />
+  );
+};
