+++ app/src/app/dashboard/statistical-unit-count-card.tsx
@@ -1,0 +1,30 @@
+import { createClient } from "@/lib/supabase/server";
+import { DashboardCard } from "@/app/dashboard/dashboard-card";
+import { StatisticalUnitIcon } from "@/components/statistical-unit-icon";
+
+export const StatisticalUnitCountCard = async ({
+  unitType,
+  title,
+}: {
+  readonly unitType: "enterprise" | "legal_unit" | "establishment";
+  readonly title: string;
+}) => {
+  const client = createClient();
+
+  const { count, error } = await client
+    .from("statistical_unit")
+    .select("", { count: "exact" })
+    .eq("unit_type", unitType)
+    .limit(0);
+
+  await new Promise((resolve) => setTimeout(resolve, 2000));
+
+  return (
+    <DashboardCard
+      title={title}
+      icon={<StatisticalUnitIcon type={unitType} className="h-4" />}
+      text={count?.toString() ?? "-"}
+      failed={!!error}
+    />
+  );
+};
